Calm down a warning.
git-svn-id: https://svn.dealii.org/trunk@8164 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/common/scripts/make_dependencies.cc
@@ -331,7 +331,7 @@ int main (int argc, char **argv)
 
                                    // parse all arguments (except the
                                    // name of the executable itself)
-  for (unsigned int c=1; c<argc; ++c)
+  for (unsigned int c=1; c<static_cast<unsigned int>(argc); ++c)
     {
       const std::string arg = argv[c];
